# public/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CGS Games API",
    "version": "1.0.0",
    "description": "Discover community Card Game Simulator games. Public reads require no authentication. Publishing and deletion use Firebase ID tokens and a user profile with a username. API errors retain the legacy error string and add code, message, and hint. GET routes also support HEAD; OPTIONS lists allowed methods. Unknown API paths return JSON 404; unsupported methods return JSON 405 with Allow."
  },
  "servers": [
    {
      "url": "https://cgs.games"
    }
  ],
...
  "security": [],
  "paths": {
    "/api/games": {
      "get": {
        "operationId": "listGames",
        "summary": "List the newest games",
        "responses": {
          "200": {
            "description": "Up to 100 games ordered by uploadedAt descending; no pagination or query filters.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "maxItems": 100,
                  "items": {
                    "$ref": "#/components/schemas/Game"
                  }
                }
              }
            }
          },
          "405": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Allow": {
                "description": "Allowed HTTP methods.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "publishGame",
        "summary": "Publish a game from its auto-update URL",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["autoUpdateUrl"],
                "properties": {
                  "autoUpdateUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "URL serving a CGS JSON game specification, including name and optional bannerImageUrl, cardBackImageUrl, and copyright."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Game published successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishedGame"
                }
              }
            }
          },
          "400": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Allow": {
                "description": "Allowed HTTP methods.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/games/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Firestore game document ID from the catalog, not a game slug.",
          "schema": {
            "type": "string"
          }
        }
      ],
      "delete": {
        "operationId": "deleteGame",
        "summary": "Delete a game owned by the authenticated creator",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Game deleted. Existing success response is plain text.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "Game deleted successfully"
              }
            }
          },
          "400": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Allow": {
                "description": "Allowed HTTP methods.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "firebaseIdToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase Auth ID token from the signed-in user. Requires a creator username."
      },
      "warmToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Operator GATCG_WARM_TOKEN secret. Only for spoiler cache maintenance; never a Firebase ID token."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error", "code", "message", "hint"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Legacy error message retained for existing clients."
          },
          "code": {
            "type": "string",
            "example": "NOT_FOUND"
          },
          "message": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          }
        }
      },
      "PublishedGame": {
        "type": "object",
        "required": ["success", "slug"],
        "properties": {
          "success": {
            "const": true
          },
          "slug": {
            "type": "string"
          }
        }
      },
      "Game": {
        "type": "object",
        "required": [
          "id",
          "username",
          "slug",
          "name",
          "bannerImageUrl",
          "autoUpdateUrl",
          "copyright",
          "uploadedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "description": "URL slug. Public game pages are /{username}/{slug}."
          },
          "name": {
            "type": "string"
          },
          "bannerImageUrl": {
            "type": "string",
            "description": "HTTPS banner URL or site-relative fallback image path."
          },
          "autoUpdateUrl": {
            "type": "string",
            "format": "uri"
          },
          "copyright": {
            "type": "string"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}